Tiny amount of cleanup
