+++ ZGF.Fonts/AtlasDirtyRect.cs
@@ -1,0 +1,19 @@
+﻿namespace ZGF.Fonts;
+
+public readonly record struct AtlasDirtyRect
+{
+    public readonly int X;
+    public readonly int Y;
+    public readonly int Width;
+    public readonly int Height;
+
+    public AtlasDirtyRect(int x, int y, int width, int height)
+    {
+        X = x;
+        Y = y;
+        Width = width;
+        Height = height;
+    }
+
+    public bool IsEmpty => Width <= 0 || Height <= 0;
+}
